+++ src/derive_input_ext.rs
@@ -1,28 +1,54 @@
-use syn::{
-    parse_quote, punctuated::Punctuated, Attribute, DeriveInput, Meta, NestedMeta, Path, Token,
-};
+use syn::{parse_quote, punctuated::Punctuated, Attribute, DeriveInput, Meta, Path, Token};
 
-use crate::{meta_list_contains, nested_meta_to_path, util};
+use crate::util;
 
 /// Functions to make it ergonomic to work with `struct` ASTs.
 pub trait DeriveInputExt {
     /// Appends derives to the list of derives.
     ///
-    /// **Note:** This can only be used with [*attribute*] macros, and not [*derive*] macros.
+    /// **Note:** This can only be used with [*attribute*] macros, and not
+    /// [*derive*] macros.
     ///
     /// * If the `derive` attribute does not exist, one will be created.
-    /// * If the `derive` attribute exists, and there are existing `derive`s that overlap with the
-    ///   derives to append, this macro will panic with the overlapping derives.
-    /// * If the `derive` attribute exists, and there are no overlapping `derive`s, then they will
-    ///   be combined.
+    /// * If the `derive` attribute exists, and there are existing `derive`s
+    ///   that overlap with the derives to append, this macro will panic with
+    ///   the overlapping derives.
+    /// * If the `derive` attribute exists, and there are no overlapping
+    ///   `derive`s, then they will be combined.
     ///
     /// # Panics
     ///
-    /// Panics if there are existing `derive`s that overlap with the derives to append.
+    /// Panics if there are existing `derive`s that overlap with the derives to
+    /// append.
     ///
     /// [*attribute*]: <https://doc.rust-lang.org/reference/procedural-macros.html#attribute-macros>
     /// [*derive*]: <https://doc.rust-lang.org/reference/procedural-macros.html#derive-mode-macros>
-    fn append_derives(&mut self, derives: Punctuated<NestedMeta, Token![,]>);
+    fn append_derives(&mut self, derives: Punctuated<Path, Token![,]>);
+
+    /// Returns whether the type contains a given `#[namespace]` attribute.
+    ///
+    /// # Parameters
+    ///
+    /// * `namespace`: The `path()` of the first-level attribute.
+    fn contains_namespace(&self, namespace: &Path) -> bool;
+
+    /// Returns the parameter from `#[namespace(parameter)]`.
+    ///
+    /// # Parameters
+    ///
+    /// * `namespace`: The `path()` of the first-level attribute.
+    ///
+    /// # Panics
+    ///
+    /// Panics if there is more than one parameter for the tag.
+    fn namespace_parameter(&self, namespace: &Path) -> Option<Meta>;
+
+    /// Returns the parameters from `#[namespace(param1, param2, ..)]`.
+    ///
+    /// # Parameters
+    ///
+    /// * `namespace`: The `path()` of the first-level attribute.
+    fn namespace_parameters(&self, namespace: &Path) -> Vec<Meta>;
 
     /// Returns whether the type contains a given `#[namespace(tag)]` attribute.
     ///
@@ -42,7 +68,7 @@ pub trait DeriveInputExt {
     /// # Panics
     ///
     /// Panics if there is more than one parameter for the tag.
-    fn tag_parameter(&self, namespace: &Path, tag: &Path) -> Option<NestedMeta>;
+    fn tag_parameter(&self, namespace: &Path, tag: &Path) -> Option<Meta>;
 
     /// Returns the parameters from `#[namespace(tag(param1, param2, ..))]`.
     ///
@@ -50,27 +76,33 @@ pub trait DeriveInputExt {
     ///
     /// * `namespace`: The `path()` of the first-level attribute.
     /// * `tag`: The `path()` of the second-level attribute.
-    fn tag_parameters(&self, namespace: &Path, tag: &Path) -> Vec<NestedMeta>;
+    fn tag_parameters(&self, namespace: &Path, tag: &Path) -> Vec<Meta>;
 }
 
 impl DeriveInputExt for DeriveInput {
-    fn append_derives(&mut self, derives_to_append: Punctuated<NestedMeta, Token![,]>) {
+    fn append_derives(&mut self, derives_to_append: Punctuated<Path, Token![,]>) {
         let attr_derives_existing = self
             .attrs
             .iter_mut()
-            .filter(|attr| attr.path.is_ident("derive"))
-            .filter_map(|attr| match attr.parse_meta() {
-                Ok(Meta::List(meta_list)) => Some((attr, meta_list)),
-                _ => None, // kcov-ignore
+            .filter(|attr| attr.path().is_ident("derive"))
+            .filter_map(|attr| {
+                match attr.parse_args_with(Punctuated::<Path, Token![,]>::parse_terminated) {
+                    Ok(derives_existing) => Some((attr, derives_existing)),
+                    _ => None, // kcov-ignore
+                }
             })
             .next();
 
         if let Some((attr, mut derives_existing)) = attr_derives_existing {
-            // Emit warning if the user derives any of the existing derives, as we do that for them.
+            // Emit warning if the user derives any of the existing derives, as we do that
+            // for them.
             let superfluous = derives_to_append
                 .iter()
-                .filter(|derive_to_append| meta_list_contains(&derives_existing, derive_to_append))
-                .filter_map(nested_meta_to_path)
+                .filter(|derive_to_append| {
+                    derives_existing
+                        .iter()
+                        .any(|derive_existing| derive_existing == *derive_to_append)
+                })
                 .map(util::format_path)
                 .collect::<Vec<_>>();
             if !superfluous.is_empty() {
@@ -78,8 +110,8 @@ impl DeriveInputExt for DeriveInput {
                 // derives_existing
                 //     .span()
                 //     .warning(
-                //         "The following are automatically derived by this proc macro attribute.",
-                //     )
+                //         "The following are automatically derived by this proc macro
+                // attribute.",     )
                 //     .emit();
                 panic!(
                     "The following are automatically derived when this attribute is used:\n\
@@ -87,14 +119,13 @@ impl DeriveInputExt for DeriveInput {
                     superfluous
                 );
             } else {
-                // derives_existing.nested.push_punct(<Token![,]>::default());
-                derives_existing.nested.extend(derives_to_append);
+                derives_existing.extend(derives_to_append);
 
                 // Replace the existing `Attribute`.
                 //
                 // `attr.parse_meta()` returns a `Meta`, which is not referenced by the
                 // `DeriveInput`, so we have to replace `attr` itself.
-                *attr = parse_quote!(#[#derives_existing]);
+                *attr = parse_quote!(#[derive(#derives_existing)]);
             }
         } else {
             // Add a new `#[derive(..)]` attribute with all the derives.
@@ -103,15 +134,27 @@ impl DeriveInputExt for DeriveInput {
         }
     }
 
+    fn contains_namespace(&self, namespace: &Path) -> bool {
+        util::contains_namespace(&self.attrs, namespace)
+    }
+
+    fn namespace_parameter(&self, namespace: &Path) -> Option<Meta> {
+        util::namespace_parameter(&self.attrs, namespace)
+    }
+
+    fn namespace_parameters(&self, namespace: &Path) -> Vec<Meta> {
+        util::namespace_parameters(&self.attrs, namespace)
+    }
+
     fn contains_tag(&self, namespace: &Path, tag: &Path) -> bool {
         util::contains_tag(&self.attrs, namespace, tag)
     }
 
-    fn tag_parameter(&self, namespace: &Path, tag: &Path) -> Option<NestedMeta> {
+    fn tag_parameter(&self, namespace: &Path, tag: &Path) -> Option<Meta> {
         util::tag_parameter(&self.attrs, namespace, tag)
     }
 
-    fn tag_parameters(&self, namespace: &Path, tag: &Path) -> Vec<NestedMeta> {
+    fn tag_parameters(&self, namespace: &Path, tag: &Path) -> Vec<Meta> {
         util::tag_parameters(&self.attrs, namespace, tag)
     }
 }
@@ -120,7 +163,7 @@ impl DeriveInputExt for DeriveInput {
 mod tests {
     use proc_macro2::Span;
     use quote::quote;
-    use syn::{parse_quote, DeriveInput, Error, Lit, Meta, NestedMeta};
+    use syn::{parse_quote, DeriveInput, Error, Meta, MetaNameValue};
 
     use super::DeriveInputExt;
 
@@ -173,6 +216,112 @@ mod tests {
     }
 
     #[test]
+    fn contains_namespace_returns_false_when_namespace_does_not_exist() -> Result<(), Error> {
+        let tokens_list = vec![
+            quote! {
+                #[other::my::derive]
+                struct Struct;
+            },
+            quote! {
+                #[my::derive::other(other)]
+                struct Struct;
+            },
+            quote! {
+                #[other(tag::name)]
+                struct Struct;
+            },
+        ];
+
+        tokens_list
+            .into_iter()
+            .try_for_each(|tokens| -> Result<(), Error> {
+                let message = format!("Failed to parse tokens: `{}`", &tokens);
+                let assertion_message = format!(
+                    "Expected `contains_namespace` to return false for tokens: `{}`",
+                    &tokens
+                );
+
+                let ast: DeriveInput =
+                    syn::parse2(tokens).map_err(|_| Error::new(Span::call_site(), &message))?;
+
+                assert!(
+                    !ast.contains_namespace(&parse_quote!(my::derive)),
+                    "{}",
+                    assertion_message
+                );
+
+                Ok(())
+            })
+    }
+
+    #[test]
+    fn namespace_parameter_returns_none_when_not_present() {
+        let ast: DeriveInput = parse_quote!(
+            #[my::derive]
+            struct Struct;
+        );
+
+        assert_eq!(ast.namespace_parameter(&parse_quote!(my::derive)), None);
+    }
+
+    #[test]
+    fn namespace_parameter_returns_ident_when_present() {
+        let ast: DeriveInput = parse_quote!(
+            #[my::derive(Magic)]
+            struct Struct;
+        );
+
+        let tag_expected: Meta = parse_quote!(Magic);
+        assert_eq!(
+            Some(tag_expected),
+            ast.namespace_parameter(&parse_quote!(my::derive))
+        );
+    }
+
+    #[test]
+    #[should_panic(expected = "Expected exactly one parameter for `#[my::derive(..)]`.")]
+    fn namespace_parameter_panics_when_multiple_parameters_present() {
+        let ast: DeriveInput = parse_quote!(
+            #[my::derive(Magic::One, Magic::Two)]
+            struct Struct;
+        );
+
+        ast.namespace_parameter(&parse_quote!(my::derive));
+    }
+
+    #[test]
+    fn namespace_parameters_returns_empty_vec_when_not_present() {
+        let ast: DeriveInput = parse_quote!(
+            struct Struct;
+        );
+
+        assert_eq!(
+            Vec::<Meta>::new(),
+            ast.namespace_parameters(&parse_quote!(my::derive))
+        );
+    }
+
+    #[test]
+    fn namespace_parameters_returns_idents_when_present() {
+        let ast: DeriveInput = parse_quote!(
+            #[my::derive(Magic::One, second = "{ Magic::Two }")]
+            struct Struct;
+        );
+
+        assert_eq!(
+            ast.namespace_parameters(&parse_quote!(my::derive)),
+            vec![
+                Meta::Path(parse_quote!(Magic::One)),
+                Meta::NameValue(MetaNameValue {
+                    path: parse_quote!(second),
+                    eq_token: Default::default(),
+                    value: parse_quote!("{ Magic::Two }")
+                }),
+            ]
+        );
+    }
+
+    #[test]
     fn contains_tag_returns_true_when_tag_exists() -> Result<(), Error> {
         let ast: DeriveInput = parse_quote!(
             #[my::derive(tag::name)]
@@ -213,12 +362,10 @@ mod tests {
                 let ast: DeriveInput =
                     syn::parse2(tokens).map_err(|_| Error::new(Span::call_site(), &message))?;
 
-                // kcov-ignore-start
                 assert!(
-                    // kcov-ignore-end
                     !ast.contains_tag(&parse_quote!(my::derive), &parse_quote!(tag::name)),
-                    "{}",              // kcov-ignore
-                    assertion_message  // kcov-ignore
+                    "{}",
+                    assertion_message
                 );
 
                 Ok(())
@@ -247,16 +394,15 @@ mod tests {
             struct Struct;
         );
 
+        let tag_expected: Meta = parse_quote!(Magic);
         assert_eq!(
-            ast.tag_parameter(&parse_quote!(my::derive), &parse_quote!(tag::name)),
-            Some(NestedMeta::Meta(Meta::Path(parse_quote!(Magic))))
+            Some(tag_expected),
+            ast.tag_parameter(&parse_quote!(my::derive), &parse_quote!(tag::name))
         );
     }
 
     #[test]
-    #[should_panic(
-        expected = "Expected exactly one identifier for `#[my::derive(tag::name(..))]`."
-    )]
+    #[should_panic(expected = "Expected exactly one parameter for `#[my::derive(tag::name(..))]`.")]
     fn tag_parameter_panics_when_multiple_parameters_present() {
         let ast: DeriveInput = parse_quote!(
             #[my::derive(tag::name(Magic::One, Magic::Two))]
@@ -274,23 +420,27 @@ mod tests {
         );
 
         assert_eq!(
-            ast.tag_parameters(&parse_quote!(my::derive), &parse_quote!(tag::name)),
-            Vec::<NestedMeta>::new()
+            Vec::<Meta>::new(),
+            ast.tag_parameters(&parse_quote!(my::derive), &parse_quote!(tag::name))
         );
     }
 
     #[test]
     fn tag_parameters_returns_idents_when_present() {
         let ast: DeriveInput = parse_quote!(
-            #[my::derive(tag::name(Magic::One, "{ Magic::Two }"))]
+            #[my::derive(tag::name(Magic::One, second = "{ Magic::Two }"))]
             struct Struct;
         );
 
         assert_eq!(
             ast.tag_parameters(&parse_quote!(my::derive), &parse_quote!(tag::name)),
             vec![
-                NestedMeta::Meta(Meta::Path(parse_quote!(Magic::One))),
-                NestedMeta::Lit(Lit::Str(parse_quote!("{ Magic::Two }"))),
+                Meta::Path(parse_quote!(Magic::One)),
+                Meta::NameValue(MetaNameValue {
+                    path: parse_quote!(second),
+                    eq_token: Default::default(),
+                    value: parse_quote!("{ Magic::Two }")
+                }),
             ]
         );
     }