+++ src/ident_ext.rs
@@ -1,5 +1,4 @@
-use std::fmt::Display;
-
+use quote::format_ident;
 use syn::Ident;
 
 /// Convenience methods on `Ident`s.
@@ -11,33 +10,32 @@ pub trait IdentExt {
     /// * `suffix`: Suffix to append.
     fn append<S>(&self, suffix: S) -> Ident
     where
-        S: Display;
+        S: quote::IdentFragment;
 
-    /// Returns a new `Ident` by prepending this Ident with the specified prefix.
+    /// Returns a new `Ident` by prepending this Ident with the specified
+    /// prefix.
     ///
     /// # Parameters
     ///
     /// * `prefix`: Prefix to prepend.
     fn prepend<S>(&self, prefix: S) -> Ident
     where
-        S: Display;
+        S: quote::IdentFragment;
 }
 
 impl IdentExt for Ident {
     fn append<S>(&self, suffix: S) -> Ident
     where
-        S: Display,
+        S: quote::IdentFragment,
     {
-        let appended = format!("{}{}", self, suffix);
-        Ident::new(&appended, self.span())
+        format_ident!("{}{}", self, suffix)
     }
 
     fn prepend<S>(&self, suffix: S) -> Ident
     where
-        S: Display,
+        S: quote::IdentFragment,
     {
-        let prepended = format!("{}{}", suffix, self);
-        Ident::new(&prepended, self.span())
+        format_ident!("{}{}", suffix, self)
     }
 }
 
@@ -68,7 +66,7 @@ mod tests {
         let one = Ident::new("One", Span::call_site());
         let two = Ident::new("Two", Span::call_site());
 
-        assert_eq!(Ident::new("OneTwo", Span::call_site()), one.append(&two));
+        assert_eq!(Ident::new("OneTwo", Span::call_site()), one.append(two));
     }
 
     #[test]
@@ -91,6 +89,6 @@ mod tests {
         let one = Ident::new("One", Span::call_site());
         let two = Ident::new("Two", Span::call_site());
 
-        assert_eq!(Ident::new("TwoOne", Span::call_site()), one.prepend(&two));
+        assert_eq!(Ident::new("TwoOne", Span::call_site()), one.prepend(two));
     }
 }